# spec/api-openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateActionOut": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "status"
        ],
        "title": "CandidateActionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryCandidateOut": {
        "properties": {
          "candidate_id": {
            "title": "Candidate Id",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "review": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review"
          },
          "source_chapter": {
            "title": "Source Chapter",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "candidate_id",
          "kind",
          "source_chapter",
          "confidence",
          "status"
        ],
        "title": "MemoryCandidateOut",
        "type": "object"
      },
...
      "RejectCandidateIn": {
        "properties": {
          "mode": {
            "default": "revise",
            "enum": [
              "revise",
              "memory_only"
            ],
            "title": "Mode",
            "type": "string"
          },
          "reason": {
            "default": "",
            "maxLength": 2000,
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "RejectCandidateIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Myink 内部 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/projects/{project_id}/candidates": {
      "get": {
        "description": "待确认池列表（按章排序，含 payload 供前端展示）。",
        "operationId": "list_candidates_internal_v1_projects__project_id__candidates_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemoryCandidateOut"
                  },
                  "title": "Response List Candidates Internal V1 Projects  Project Id  Candidates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Candidates",
        "tags": [
          "candidates"
        ]
      }
    },
    "/internal/v1/projects/{project_id}/candidates/{candidate_id}/confirm": {
      "post": {
        "description": "确认候选落库（Event/Fact/CharacterState…，§7.3）。幂等：已处理候选返回 409。",
        "operationId": "confirm_candidate_internal_v1_projects__project_id__candidates__candidate_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
    "/internal/v1/projects/{project_id}/candidates/{candidate_id}/reject": {
      "post": {
        "description": "拒绝候选（幻觉/误抽清理）。幂等：已处理候选返回 409。",
        "operationId": "reject_candidate_internal_v1_projects__project_id__candidates__candidate_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RejectCandidateIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
...
  }
}